Replaced the "make -j11" hard-coded values with a call to a new helper script, tools/mmake which calls another newly-added script (tools/corecount) to get the core count on this system and multiply it by 1.5 and pass it to make -j.  It currently knows how to do this on Linux and Mac OS X.  For all other systems, it returns 1, so it gives you make -j1, since the math truncates, rather than rounds.
This is not just a nicety: "make -j11" can lock up a Pi in some cases.
It appears to be worst when most of the files in the system have to be
rebuilt.  It's probably running the system out of RAM, and it isn't
handling it gracefully.

FossilOrigin-Name: cffeb65eb80a14202e466e6063d9376efef8a47d