Factored the vast bulk of the PiDP-8/I specific code out of PDP8/pdp8_cpu.c into a set of several new functions defined in a new PDP8/pidp8i.c module.  Along the way, reworked the code style to match the rest of SIMH.
Very little is functionally different, and the differences are intended
to only be spot improvements not needing a separate checkin comment.
Examples: a) some of the bit twiddling math was overly complex, so
reduced several 3-ish line calculations to one-liners; appended
strerror(errno) to some error message strings; c) reorganized the switch
testing code a bit to reduce the number of times certain tests and
calculations are made.

This change also greatly reduces the number of PiDP-8/I specific
globals.  All but one of the app-wide globals that used to be defined in
pdp8_cpu.c.in are now scoped to the pidp8i.c module, and all but one of
those are statics scoped to a single function in that module.

Because this refactoring moves the @BOOTDIR@ usage from pdp8_cpu.c to
the new pidp8i module, the SIMH module is back to being a static C file,
while pdp8i.c is generated from pidp8i.c.in.

Also added a whole bunch more comments, and moved some prior comments
into bug tickets.

Checking this in as a branch because it is so extensive that I might
have broken something.  It hasn't been tested yet because I did the work
on my fast desktop PC, not on the Pi, there being no point since the
changes are so wide-reaching that it didn't even start building again
until just before this checkin.  Now it's time to do that testing.

FossilOrigin-Name: 642a7201a017bcf03f3be900a389d97070ce2db2