Fixed a couple of problems in the way SING INST and CONT were handled. These are related because in the old way of doing things, these two switches had a kind of lock-step operation that tried to split instruction execution over two separate passes of the PDP-8 sim's CPU instruction decoding loop.  Now these two switches are handled more or less the same way as all the others, in a single cycle of the instruction loop.
With the prior scheme, while the processor was stopped with a HLT
instruction, pressing CONT would basically jump 2 instructions forward,
rather than continue to the instruction after the HLT, skipping it.
This fix closes [f961906a5c].

While in there and reading the DEC documentation describing the proper
operation of the switches, I found that we weren't handling SING_INST
properly at all.  When flipped with the processor in free-running mode,
it is supposed to stop the processor just as if you'd pressed STOP.
Now it does.  (I could have checked this in separately, but the two
fixes are in overlapping and interrelated sections of code, so it
doesn't make sense to check in a purposely-written erroneous SING INST
handler just to get a separate checkin comment.)

FossilOrigin-Name: 9a1277327070e5abb8ffd6ecea5b9d32873e7782