Removed the code that skips updating the global ledstatus[] values from the PDP-8 CPU core based on the human PoV limit.  Because the GPIO thread runs asynchronously with it and has its own timing, these two high-jitter timing loops created a messy beat frequency on the front panel, which you could see in a tight KSF ; JMP .-1 loop, as OS/8 does when idle, waiting for a key.
This mechanism came to us in the merger of the common parts of ILS and
NLS, and ILS only needed it when it was modifying the LED brightness
values values in the SIMH thread instead of in the GPIO thread as it
does now.  The only argument for continuing to do this — and for also
doing it for NLS — is that it avoids a bunch of recalculation that
the human couldn't see, but that is only true if the GPIO thread is
able to stochastically sample the LED states so that even with the
asynchronous jitter going on, a tight loop like the above eventually
blurs together to a nice 50/50 mix of the two instructions' LED states,
especially under the ILS.

Ultimately, any attention paid to the human PoV limit has to be over
in the GPIO thread.

FossilOrigin-Name: 61ce7d6852d4e684990beeeb7edfbea7f1c13144