Moved the @abs_top_srcdir@ and @builddir@ stuff out of libexec/mkos8.in into lib/mkos8/dirs.py.in, which exists only to hold those directory definitions.  Then reworked the make dependencies so that simply regenerating the lib/mkos8/dirs.py file no longer causes a rebuild of the OS/8 media, since by far the most likely reason for this is that auto.def or some other *.in file changed.
Also removed the simulator as a make dependency of the OS/8 RK05 media,
since we just need it to *exist*, not to be older than the media.
Instead, we check for it before rebuilding the media, and we only do
that when one of the input files for building the media change, which
explicitly does not include dirs.py.

There's one missing step in all of this which can cause us to miss a
needed rebuild, but that will be taken care of next.

FossilOrigin-Name: 9b76afeacb834e8d05f606ed29c2facc3f050a50